# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Internbot API",
    "version": "1.0.0",
    "description": "Backend workflow API for the Internbot internship platform. Source of truth: [WORKFLOW-API-SPEC.md](https://github.com/giatinhuynh/Internbot/blob/develop/docs/WORKFLOW-API-SPEC.md)."
  },
...
  "paths": {
...
    "/api/v1/internships": {
      "get": {
        "operationId": "listInternships",
        "summary": "List internships visible to the caller",
        "description": "Students are server-filtered to their own internships. Coordinators can list all internships and filter by status, opportunity, or user. See WORKFLOW-API-SPEC.md §7.4.",
        "tags": ["Internships"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "applied",
                  "offer_pending_review",
                  "offer_changes_requested",
                  "offer_approved",
                  "rejected"
                ]
              }
            }
          },
          {
            "in": "query",
            "name": "opportunityId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "in": "query",
            "name": "pageToken",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "description": "Default `-createdAt`. Prefix with `-` for descending.",
              "type": "string",
              "enum": ["createdAt", "-createdAt", "lastSubmittedAt", "-lastSubmittedAt"]
            },
            "description": "Default `-createdAt`. Prefix with `-` for descending."
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of internships.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternshipListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query string or student attempted to widen user filter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Firebase ID token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createInternship",
        "summary": "Apply to an opportunity",
        "description": "Student-only. Creates an internship application, apply activity, and coordinator notifications.",
        "tags": ["Internships"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInternshipRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Internship created.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternshipResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not a student.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Opportunity does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "No selected semester, opportunity unpublished, semester mismatch, or duplicate application.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Missing required fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internships/{id}": {
      "get": {
        "operationId": "getInternship",
        "summary": "Return an internship by id",
        "description": "Student owner or coordinator can read an internship.",
        "tags": ["Internships"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "int_042",
              "description": "Platform internship id."
            },
            "required": true,
            "description": "Platform internship id."
          }
        ],
        "responses": {
          "200": {
            "description": "Internship found.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternshipResponse"
                }
              }
            }
          },
          "403": {
            "description": "Student cannot see this internship.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No internship exists with the supplied id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "patchInternship",
        "summary": "Update internship offer details",
        "description": "Student-owner only. Coordinators receive 405 and should use decisions or comments instead.",
        "tags": ["Internships"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "int_042",
              "description": "Platform internship id."
            },
            "required": true,
            "description": "Platform internship id."
          },
          {
            "in": "header",
            "name": "if-match",
            "schema": {
              "description": "Opt-in optimistic concurrency — the current ETag from a prior GET.",
              "type": "string"
            },
            "description": "Opt-in optimistic concurrency — the current ETag from a prior GET."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchInternshipRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Internship updated.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternshipResponse"
                }
              }
            }
          },
          "400": {
            "description": "Body contains immutable or unknown fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Student caller does not own the internship.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "Coordinator callers cannot PATCH internships.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Internship is terminal and not editable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "412": {
            "description": "Stale `If-Match`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Empty body or domain validation failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internships/{id}/attachments/{attachmentId}": {
      "get": {
        "operationId": "getInternshipAttachment",
        "summary": "Return an internship attachment download resource",
        "description": "Returns attachment metadata with a fresh short-lived V4 signed Cloud Storage URL. Students can read only their own internship attachments; coordinators can read all.",
        "tags": ["Internships"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "int_042",
              "description": "Platform internship id."
            },
            "required": true,
            "description": "Platform internship id."
          },
          {
            "in": "path",
            "name": "attachmentId",
            "schema": {
              "type": "string",
              "example": "att_101",
              "description": "Attachment id."
            },
            "required": true,
            "description": "Attachment id."
          }
        ],
        "responses": {
          "200": {
            "description": "Attachment found with a fresh signed download URL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternshipAttachmentDownloadResponse"
                }
              }
            }
          },
          "403": {
            "description": "Student caller does not own the parent internship.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No internship or attachment exists with the supplied id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internships/{id}/offer-submissions": {
      "post": {
        "operationId": "submitInternshipOffer",
        "summary": "Submit an internship offer for coordinator review",
        "description": "Student-owner only. Requires at least one offer attachment and transitions applied/changes_requested to offer_pending_review.",
        "tags": ["Internships"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "int_042",
              "description": "Platform internship id."
            },
            "required": true,
            "description": "Platform internship id."
          },
          {
            "in": "header",
            "name": "if-match",
            "schema": {
              "description": "Opt-in optimistic concurrency — the current ETag from a prior GET.",
              "type": "string"
            },
            "description": "Opt-in optimistic concurrency — the current ETag from a prior GET."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitInternshipOfferRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Offer submitted.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternshipResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not the student owner.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Internship is not in an offer-submittable state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "412": {
            "description": "Stale `If-Match`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Missing offer details or offer attachment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internships/{id}/comments": {
      "post": {
        "operationId": "addInternshipComment",
        "summary": "Add a comment to an internship timeline",
        "description": "Student owner or coordinator can comment in any internship state. Comments do not rotate the parent ETag.",
        "tags": ["Internships"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "int_042",
              "description": "Platform internship id."
            },
            "required": true,
            "description": "Platform internship id."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddInternshipCommentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment activity created.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternshipActivityResponse"
                }
              }
            }
          },
          "403": {
            "description": "Student caller does not own the internship.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No internship exists with the supplied id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Missing or empty text.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateInternshipRequest": {
        "type": "object",
        "properties": {
          "opportunityId": {
            "type": "string",
            "minLength": 1,
            "example": "opp_042"
          }
        },
        "required": ["opportunityId"],
        "additionalProperties": false,
        "description": "Body for POST /api/v1/internships. See WORKFLOW-API-SPEC.md §7.4."
      },
      "PatchInternshipRequest": {
        "type": "object",
        "properties": {
          "offerDate": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "endDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "description": "Body for PATCH /api/v1/internships/:id. Lifecycle and ownership fields are rejected with 400."
      },
      "SubmitInternshipOfferRequest": {
        "type": "object",
        "properties": {
          "offerDate": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "endDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": ["offerDate", "startDate"],
        "additionalProperties": false,
        "description": "Body for POST /api/v1/internships/:id/offer-submissions."
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "https://httpstatuses.io/403"
          },
          "title": {
            "type": "string",
            "example": "Forbidden"
          },
          "status": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "detail": {
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": ["type", "title", "status", "detail", "error"],
        "additionalProperties": false,
        "description": "RFC 9457 Problem Details wrapped with the domain error envelope — see docs/ERROR-HANDLING.md."
      },
...
      "InternshipListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InternshipListItemResponse"
            }
          },
          "nextPageToken": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": ["items", "nextPageToken"],
        "additionalProperties": false,
        "description": "Paginated list of internships."
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token",
        "description": "Firebase ID token obtained client-side. Platform identity is hydrated at the api edge from `userIdentities/{provider}__{uid}` → `users/{id}` on every request; on a brand-new student-shape email the hydrator JIT-creates the platform record in the same transaction. See docs/WORKFLOW-API-SPEC.md §7.0."
      }
    }
  }
}